make it so the optimisation is based off the total items of the target type left over, not the total items of the target type in flow
this one has quite a funny story behind it.  basically, i saw that i was optimising based on the items in flow, and decided to create a "virtual recipe" for the target item that would consume one of it per minute and produce nothing (but it is of the item type rather than of the recipe type to differentiate it from an actual recipe).  i would then optimise using the virtual recipe instead of the item.  what i then tested was 10 iron ore and 1 sam ore, producing iron ore.  i got more than 10, suggesting that the program was still somehow optimising based on the items in flow.  much debugging later, i couldnt see any issues.  working through the recipes it ended up using by hand, i found that the algorithm had discovered a way to turn one sam ore into 9/550 iron ore, as long as it could be jumpstarted by an input of coal, sulphur, or limestone, and sufficient power could be supplied (the problem does not account for power yet, and the program permits resource loops since it has no effective means to detect them, and they are used in high-efficiency production of oil products with the recycled rubber and plastic recipes converting fuel into more rubber and plastic, cycling some plastic and rubber within the system)
it also turned out that after reverting my changes and using the same test input of 10 iron ore and 1 sam ore, the program actually did optimise for the amount of iron ore being transported, not the amount of iron ore available at the end.  so i fixed a bug somewhat unknowingly.